@@ -6,9 +6,10 @@ import (
 	"errors"
 	"fmt"
 	"io"
-	"log"
+	"log/slog"
 	"os"
 	"strconv"
+	"sync"
 	"time"
 
 	"github.com/Songmu/flextime"
@@ -17,10 +18,17 @@ import (
 	"github.com/aws/aws-sdk-go-v2/service/dynamodb/types"
 	"github.com/aws/smithy-go"
 	"github.com/gofrs/flock"
-	logx "github.com/mashiike/go-logx"
-	"github.com/samber/lo"
 	"github.com/shogo82148/go-retry"
 )
+
+type StorageOption struct {
+	Type             string `help:"storage type" default:"dynamodb" enum:"dynamodb,file" env:"GDNOTIFY_STORAGE_TYPE"`
+	TableName        string `help:"dynamodb table name" default:"gdnotify" env:"GDNOTIFY_DDB_TABLE_NAME"`
+	AutoCreate       bool   `help:"auto create dynamodb table" default:"false" env:"GDNOTIFY_DDB_AUTO_CREATE" negatable:""`
+	DynamoDBEndpoint string `help:"dynamodb endpoint" env:"GDNOTIFY_DDB_ENDPOINT"`
+	DataFile         string `help:"file storage data file" default:"gdnotify.dat" env:"GDNOTIFY_FILE_STORAGE_DATA_FILE"`
+	LockFile         string `help:"file storage lock file" default:"gdnotify.lock" env:"GDNOTIFY_FILE_STORAGE_LOCK_FILE"`
+}
 
 type ChannelItem struct {
 	ChannelID          string
@@ -36,8 +44,13 @@ type ChannelItem struct {
 func (item *ChannelItem) IsAboutToExpired(ctx context.Context, remaining time.Duration) bool {
 	now := flextime.Now()
 	d := item.Expiration.Sub(now)
-	logx.Printf(ctx, "[debug] IsAboutToExpired remaining=%s expiration=%s, now=%s, channel_id=%s, resource_id=%s, drive_id=%s ",
-		d, item.Expiration.Format(time.RFC3339), now.Format(time.RFC3339), item.ChannelID, item.ResourceID, item.DriveID,
+	slog.DebugContext(ctx, "IsAboutToExpired",
+		"remaining", remaining,
+		"expiration", item.Expiration.Format(time.RFC3339),
+		"now", now.Format(time.RFC3339),
+		"channel_id", item.ChannelID,
+		"resource_id", item.ResourceID,
+		"drive_id", item.DriveID,
 	)
 	return d <= remaining
 }
@@ -141,11 +154,11 @@ type Storage interface {
 	DeleteChannel(context.Context, *ChannelItem) error
 }
 
-type ChannelNotFound struct {
+type ChannelNotFoundError struct {
 	ChannelID string
 }
 
-func (err *ChannelNotFound) Error() string {
+func (err *ChannelNotFoundError) Error() string {
 	return fmt.Sprintf("channel_id:%s not found", err.ChannelID)
 }
 
@@ -157,14 +170,14 @@ func (err *ChannelAlreadyExists) Error() string {
 	return fmt.Sprintf("channel_id:%s already exists", err.ChannelID)
 }
 
-func NewStorage(ctx context.Context, cfg *StorageConfig, awsCfg aws.Config) (Storage, func() error, error) {
+func NewStorage(ctx context.Context, cfg StorageOption) (Storage, error) {
 	switch cfg.Type {
-	case StorageTypeDynamoDB:
-		return NewDynamoDBStorage(ctx, cfg, awsCfg)
-	case StorageTypeFile:
+	case "dynamodb":
+		return NewDynamoDBStorage(ctx, cfg)
+	case "file":
 		return NewFileStorage(ctx, cfg)
 	}
-	return nil, nil, errors.New("unknown storage type")
+	return nil, errors.New("unknown storage type")
 }
 
 type DynamoDBStorage struct {
@@ -172,27 +185,37 @@ type DynamoDBStorage struct {
 	tableName string
 }
 
-func NewDynamoDBStorage(ctx context.Context, cfg *StorageConfig, awsCfg aws.Config) (*DynamoDBStorage, func() error, error) {
-	s := &DynamoDBStorage{
-		client:    dynamodb.NewFromConfig(awsCfg),
-		tableName: *cfg.TableName,
+func NewDynamoDBStorage(ctx context.Context, cfg StorageOption) (*DynamoDBStorage, error) {
+	awsCfg, err := loadAWSConfig()
+	if err != nil {
+		return nil, err
 	}
-	logx.Printf(ctx, "[info] check describe dynamodb table `%s`", s.tableName)
+	opts := []func(*dynamodb.Options){}
+	if cfg.DynamoDBEndpoint != "" {
+		opts = append(opts, func(o *dynamodb.Options) {
+			o.BaseEndpoint = aws.String(cfg.DynamoDBEndpoint)
+		})
+	}
+	s := &DynamoDBStorage{
+		client:    dynamodb.NewFromConfig(awsCfg, opts...),
+		tableName: cfg.TableName,
+	}
+	slog.InfoContext(ctx, "check describe dynamodb table", "table_name", s.tableName)
 	exists, err := s.tableExists(ctx)
 	if err != nil {
-		return nil, nil, err
+		return nil, err
 	}
-	if !exists {
+	if !exists && cfg.AutoCreate {
 		if err := s.createTable(ctx); err != nil {
-			return nil, nil, err
+			return nil, err
 		}
 	}
 
-	return s, nil, nil
+	return s, nil
 }
 
 func (s *DynamoDBStorage) tableExists(ctx context.Context) (bool, error) {
-	logx.Printf(ctx, "[debug] check describe dynamodb table `%s`", s.tableName)
+	slog.DebugContext(ctx, "check describe dynamodb table", "table_name", s.tableName)
 	table, err := s.client.DescribeTable(ctx, &dynamodb.DescribeTableInput{
 		TableName: aws.String(s.tableName),
 	})
@@ -203,10 +226,10 @@ func (s *DynamoDBStorage) tableExists(ctx context.Context) (bool, error) {
 				return false, nil
 			}
 		}
-		logx.Println(ctx, "[debug] DescribeTable: ", err)
+		slog.DebugContext(ctx, "DescribeTable", "error", err)
 		return false, err
 	}
-	logx.Printf(ctx, "[debug] exists table `%s` status is `%s`", s.tableName, table.Table.TableStatus)
+	slog.DebugContext(ctx, "exists table", "table_name", s.tableName, "status", table.Table.TableStatus)
 	if table.Table.TableStatus == types.TableStatusActive || table.Table.TableStatus == types.TableStatusUpdating {
 		return true, nil
 	}
@@ -224,14 +247,14 @@ func (s *DynamoDBStorage) waitTableActive(ctx context.Context) error {
 	retrier := policy.Start(ctx)
 	var err error
 	var exists bool
-	logx.Printf(ctx, "[debug] start wait dynamodb table `%s` active", s.tableName)
+	slog.DebugContext(ctx, "start wait dynamodb table active", "table_name", s.tableName)
 	for retrier.Continue() {
 		exists, err = s.tableExists(ctx)
 		if err == nil && exists {
 			return nil
 		}
 	}
-	logx.Printf(ctx, "[debug] timeout wait dynamodb table `%s` active", s.tableName)
+	slog.DebugContext(ctx, "timeout wait dynamodb table active", "table_name", s.tableName)
 	if err == nil {
 		return fmt.Errorf("table not active")
 	}
@@ -239,7 +262,7 @@ func (s *DynamoDBStorage) waitTableActive(ctx context.Context) error {
 }
 
 func (s *DynamoDBStorage) createTable(ctx context.Context) error {
-	logx.Printf(ctx, "[debug] create dynamodb table `%s`", s.tableName)
+	slog.DebugContext(ctx, "create dynamodb table", "table_name", s.tableName)
 	output, err := s.client.CreateTable(ctx, &dynamodb.CreateTableInput{
 		TableName: aws.String(s.tableName),
 		AttributeDefinitions: []types.AttributeDefinition{
@@ -260,17 +283,17 @@ func (s *DynamoDBStorage) createTable(ctx context.Context) error {
 		var ae smithy.APIError
 		if errors.As(err, &ae) {
 			if ae.ErrorCode() == "ResourceInUseException" {
-				logx.Printf(ctx, "[debug] crate dynamodb table `%s` ResourceInUseException: wait table active", s.tableName)
+				slog.DebugContext(ctx, "crate dynamodb table ResourceInUseException: wait table active", "table_name", s.tableName)
 				if err := s.waitTableActive(ctx); err != nil {
 					return err
 				}
 				return nil
 			}
 		}
-		logx.Println(ctx, "[debug] CreateTable failed: ", err)
+		slog.DebugContext(ctx, "CreateTable failed", "error", err)
 		return err
 	}
-	logx.Printf(ctx, "[info] create dynamodb table `%s`", *output.TableDescription.TableArn)
+	slog.InfoContext(ctx, "create dynamodb table", "table_arn", *output.TableDescription.TableArn)
 	if err := s.waitTableActive(ctx); err != nil {
 		return err
 	}
@@ -278,31 +301,31 @@ func (s *DynamoDBStorage) createTable(ctx context.Context) error {
 }
 
 func (s *DynamoDBStorage) FindAllChannels(ctx context.Context) (<-chan []*ChannelItem, error) {
-	logx.Printf(ctx, "[debug] scan dynamodb table `%s`", s.tableName)
+	slog.DebugContext(ctx, "scan dynamodb table", "table_name", s.tableName)
 	output, err := s.client.Scan(ctx, &dynamodb.ScanInput{
 		TableName:      aws.String(s.tableName),
 		Select:         types.SelectAllAttributes,
 		ConsistentRead: aws.Bool(false),
 	})
 	if err != nil {
-		logx.Printf(ctx, "[debug] scan dynamodb table failed: %s", err.Error())
+		slog.DebugContext(ctx, "scan dynamodb table failed", "error", err)
 		return nil, err
 	}
-	logx.Printf(ctx, "[debug] scan dynamodb table success item_count=%d", output.Count)
+	slog.DebugContext(ctx, "scan dynamodb table success", "item_count", output.Count)
 	ch := make(chan []*ChannelItem, 10)
-	ch <- lo.Map(output.Items, func(values map[string]types.AttributeValue, _ int) *ChannelItem {
+	ch <- Map(output.Items, func(values map[string]types.AttributeValue) *ChannelItem {
 		return NewChannelItemWithDynamoDBAttributeValues(values)
 	})
 	if output.LastEvaluatedKey == nil {
-		logx.Printf(ctx, "[debug] LastEvaluatedKey is null return FindAllChannels")
+		slog.DebugContext(ctx, "LastEvaluatedKey is null return FindAllChannels")
 		close(ch)
 		return ch, nil
 	}
-	logx.Printf(ctx, "[debug] need background scan dynamodb table")
+	slog.DebugContext(ctx, "need background scan dynamodb table")
 	go func() {
-		logx.Printf(ctx, "[debug] start background scan dynamodb table `%s`", s.tableName)
+		slog.DebugContext(ctx, "start background scan dynamodb table", "table_name", s.tableName)
 		defer func() {
-			logx.Printf(ctx, "[debug] finish background scan dynamodb table `%s`", s.tableName)
+			slog.DebugContext(ctx, "finish background scan dynamodb table", "table_name", s.tableName)
 			close(ch)
 		}()
 		for output.LastEvaluatedKey != nil {
@@ -312,11 +335,11 @@ func (s *DynamoDBStorage) FindAllChannels(ctx context.Context) (<-chan []*Channe
 				ConsistentRead: aws.Bool(false),
 			})
 			if err != nil {
-				logx.Printf(ctx, "[error] background scan dynamodb table failed: %s", err.Error())
+				slog.ErrorContext(ctx, "background scan dynamodb table failed", "error", err)
 				return
 			}
-			logx.Printf(ctx, "[debug] background scan dynamodb table success item_count=%d", output.Count)
-			ch <- lo.Map(output.Items, func(values map[string]types.AttributeValue, _ int) *ChannelItem {
+			slog.DebugContext(ctx, "background scan dynamodb table success", "item_count", output.Count)
+			ch <- Map(output.Items, func(values map[string]types.AttributeValue) *ChannelItem {
 				return NewChannelItemWithDynamoDBAttributeValues(values)
 			})
 			time.Sleep(100 * time.Millisecond)
@@ -326,7 +349,7 @@ func (s *DynamoDBStorage) FindAllChannels(ctx context.Context) (<-chan []*Channe
 }
 
 func (s *DynamoDBStorage) SaveChannel(ctx context.Context, item *ChannelItem) error {
-	logx.Printf(ctx, "[debug] put item channel_id=`%s` to dynamodb table `%s`", item.ChannelID, s.tableName)
+	slog.DebugContext(ctx, "put item to dynamodb table", "channel_id", item.ChannelID, "table_name", s.tableName)
 	_, err := s.client.PutItem(ctx, &dynamodb.PutItemInput{
 		TableName:           aws.String(s.tableName),
 		Item:                item.ToDynamoDBAttributeValues(),
@@ -334,7 +357,7 @@ func (s *DynamoDBStorage) SaveChannel(ctx context.Context, item *ChannelItem) er
 	})
 	if err != nil {
 		var ae smithy.APIError
-		logx.Printf(ctx, "[warn] failed put item channel_id=`%s` resource_id=%s to dynamodb table `%s`: %s", item.ChannelID, item.ResourceID, s.tableName, err.Error())
+		slog.WarnContext(ctx, "failed put item to dynamodb table", "channel_id", item.ChannelID, "resource_id", item.ResourceID, "table_name", s.tableName, "error", err)
 		if errors.As(err, &ae) {
 			if ae.ErrorCode() == "ConditionalCheckFailedException" {
 				return &ChannelAlreadyExists{ChannelID: item.ChannelID}
@@ -342,12 +365,12 @@ func (s *DynamoDBStorage) SaveChannel(ctx context.Context, item *ChannelItem) er
 		}
 		return err
 	}
-	logx.Printf(ctx, "[info] put item channel_id=`%s` to dynamodb table `%s`", item.ChannelID, s.tableName)
+	slog.InfoContext(ctx, "put item to dynamodb table", "channel_id", item.ChannelID, "table_name", s.tableName)
 	return nil
 }
 
 func (s *DynamoDBStorage) UpdatePageToken(ctx context.Context, target *ChannelItem) error {
-	logx.Printf(ctx, "[debug] update item channel_id=`%s` to dynamodb table `%s`", target.ChannelID, s.tableName)
+	slog.DebugContext(ctx, "update item to dynamodb table", "channel_id", target.ChannelID, "table_name", s.tableName)
 	values := target.ToDynamoDBAttributeValues()
 	_, err := s.client.UpdateItem(ctx, &dynamodb.UpdateItemInput{
 		TableName: aws.String(s.tableName),
@@ -368,15 +391,15 @@ func (s *DynamoDBStorage) UpdatePageToken(ctx context.Context, target *ChannelIt
 		},
 	})
 	if err != nil {
-		logx.Printf(ctx, "[warn] failed update item channel_id=`%s` to dynamodb table `%s` page_token=%s", target.ChannelID, s.tableName, target.PageToken)
+		slog.WarnContext(ctx, "failed update item to dynamodb table", "channel_id", target.ChannelID, "table_name", s.tableName, "page_token", target.PageToken)
 		return err
 	}
-	logx.Printf(ctx, "[info] update item channel_id=`%s` to dynamodb table `%s` page_token=%s", target.ChannelID, s.tableName, target.PageToken)
+	slog.InfoContext(ctx, "update item to dynamodb table", "channel_id", target.ChannelID, "table_name", s.tableName, "page_token", target.PageToken)
 	return nil
 }
 
 func (s *DynamoDBStorage) DeleteChannel(ctx context.Context, target *ChannelItem) error {
-	logx.Printf(ctx, "[debug] delete item channel_id=`%s` from dynamodb table `%s`", target.ChannelID, s.tableName)
+	slog.DebugContext(ctx, "delete item from dynamodb table", "channel_id", target.ChannelID, "table_name", s.tableName)
 	_, err := s.client.DeleteItem(ctx, &dynamodb.DeleteItemInput{
 		TableName: aws.String(s.tableName),
 		Key: map[string]types.AttributeValue{
@@ -387,15 +410,15 @@ func (s *DynamoDBStorage) DeleteChannel(ctx context.Context, target *ChannelItem
 		ConditionExpression: aws.String("attribute_exists(ChannelID)"),
 	})
 	if err != nil {
-		logx.Printf(ctx, "[warn] failed delete item channel_id=`%s` resource_id=%s from dynamodb table `%s`", target.ChannelID, target.ResourceID, s.tableName)
+		slog.WarnContext(ctx, "failed delete item from dynamodb table", "channel_id", target.ChannelID, "resource_id", target.ResourceID, "table_name", s.tableName)
 		return err
 	}
-	logx.Printf(ctx, "[info] delete item channel_id=`%s` resource_id=`%s` from dynamodb table `%s`", target.ChannelID, target.ChannelID, s.tableName)
+	slog.InfoContext(ctx, "delete item from dynamodb table", "channel_id", target.ChannelID, "resource_id", target.ChannelID, "table_name", s.tableName)
 	return nil
 }
 
 func (s *DynamoDBStorage) FindOneByChannelID(ctx context.Context, channelID string) (*ChannelItem, error) {
-	logx.Printf(ctx, "[debug] get item channel_id=`%s` from dynamodb table `%s`", channelID, s.tableName)
+	slog.DebugContext(ctx, "get item from dynamodb table", "channel_id", channelID, "table_name", s.tableName)
 	output, err := s.client.GetItem(ctx, &dynamodb.GetItemInput{
 		TableName: aws.String(s.tableName),
 		Key: map[string]types.AttributeValue{
@@ -405,27 +428,32 @@ func (s *DynamoDBStorage) FindOneByChannelID(ctx context.Context, channelID stri
 		},
 	})
 	if err != nil {
-		logx.Printf(ctx, "[warn] failed get item channel_id=`%s` from dynamodb table `%s`", channelID, s.tableName)
+		slog.WarnContext(ctx, "get item failed", "channel_id", channelID, "table_name", s.tableName, "error", err)
 		return nil, err
 	}
-	logx.Printf(ctx, "[debug] success get item channel_id=`%s` from dynamodb table `%s`", channelID, s.tableName)
+	if output.Item == nil {
+		slog.WarnContext(ctx, "not found item", "channel_id", channelID, "table_name", s.tableName)
+		return nil, &ChannelNotFoundError{ChannelID: channelID}
+	}
+	slog.DebugContext(ctx, "success get item", "channel_id", channelID, "table_name", s.tableName)
 	return NewChannelItemWithDynamoDBAttributeValues(output.Item), nil
 }
 
 type FileStorage struct {
+	mu    sync.Mutex
 	Items []*ChannelItem
 
 	LockFile string
 	FilePath string
 }
 
-func NewFileStorage(ctx context.Context, cfg *StorageConfig) (*FileStorage, func() error, error) {
+func NewFileStorage(_ context.Context, cfg StorageOption) (*FileStorage, error) {
 	s := &FileStorage{
-		FilePath: *cfg.DataFile,
-		LockFile: *cfg.LockFile,
+		FilePath: cfg.DataFile,
+		LockFile: cfg.LockFile,
 	}
 
-	return s, nil, nil
+	return s, nil
 }
 
 func (s *FileStorage) FindAllChannels(ctx context.Context) (<-chan []*ChannelItem, error) {
@@ -435,7 +463,7 @@ func (s *FileStorage) FindAllChannels(ctx context.Context) (<-chan []*ChannelIte
 			ch <- s.Items
 			return nil
 		}); err != nil {
-			logx.Println(ctx, "[error] failed background channels read:", err)
+			slog.ErrorContext(ctx, "failed background channels read", "error", err)
 		}
 		close(ch)
 	}()
@@ -459,15 +487,13 @@ func (s *FileStorage) UpdatePageToken(ctx context.Context, target *ChannelItem) 
 	return s.transactional(ctx, func(context.Context) error {
 		for i, c := range s.Items {
 			if c.ChannelID == target.ChannelID {
-				logx.Printf(ctx, "[debug] update PageToken channel_id=%s old_page_token=%s new_page_token=%s",
-					s.Items[i].ChannelID, s.Items[i].PageToken, target.PageToken,
-				)
+				slog.DebugContext(ctx, "update PageToken", "channel_id", s.Items[i].ChannelID, "old_page_token", s.Items[i].PageToken, "new_page_token", target.PageToken)
 				s.Items[i].PageToken = target.PageToken
 
 				return nil
 			}
 		}
-		return &ChannelNotFound{ChannelID: target.ChannelID}
+		return &ChannelNotFoundError{ChannelID: target.ChannelID}
 	})
 }
 
@@ -489,24 +515,22 @@ func (s *FileStorage) FindOneByChannelID(ctx context.Context, channelID string) 
 		for _, item := range s.Items {
 			if item.ChannelID == channelID {
 				ret = item
-				logx.Printf(ctx, "[debug] found ChannelItem channel_id=%s resource_id=%s drive_id=%s:",
-					ret.ChannelID, ret.ResourceID, ret.DriveID,
-				)
+				slog.DebugContext(ctx, "found ChannelItem", "channel_id", ret.ChannelID, "resource_id", ret.ResourceID, "drive_id", ret.DriveID)
 				return nil
 			}
 		}
-		return &ChannelNotFound{ChannelID: channelID}
+		return &ChannelNotFoundError{ChannelID: channelID}
 	}); err != nil {
-		logx.Println(ctx, "[debug] failed read:", err)
+		slog.DebugContext(ctx, "failed read", "error", err)
 		return nil, err
 	}
-	logx.Printf(ctx, "[debug] return ChannelItem channel_id=%s resource_id=%s drive_id=%s:",
-		ret.ChannelID, ret.ResourceID, ret.DriveID,
-	)
+	slog.DebugContext(ctx, "return ChannelItem", "channel_id", ret.ChannelID, "resource_id", ret.ResourceID, "drive_id", ret.DriveID)
 	return ret, nil
 }
 
 func (s *FileStorage) transactional(ctx context.Context, fn func(context.Context) error) error {
+	s.mu.Lock()
+	defer s.mu.Unlock()
 	fileLock := flock.New(s.LockFile)
 	policy := retry.Policy{
 		MinDelay: 100 * time.Millisecond,
@@ -519,14 +543,14 @@ func (s *FileStorage) transactional(ctx context.Context, fn func(context.Context
 	var err error
 	var locked bool
 	for retrier.Continue() {
-		logx.Println(ctx, "[debug] try file storage lock:", s.LockFile)
+		slog.DebugContext(ctx, "try file storage lock", "lock_file", s.LockFile)
 		locked, err = fileLock.TryLock()
 		if err != nil {
-			logx.Println(ctx, "[debug] get file storage lock failed :", err)
+			slog.DebugContext(ctx, "get file storage lock failed", "error", err)
 			continue
 		}
 		if locked {
-			logx.Println(ctx, "[debug] get file storage lock success")
+			slog.DebugContext(ctx, "get file storage lock success")
 			break
 		}
 	}
@@ -535,35 +559,35 @@ func (s *FileStorage) transactional(ctx context.Context, fn func(context.Context
 	}
 	defer func() {
 		if err := fileLock.Unlock(); err != nil {
-			logx.Println(ctx, "[debug] file storage unlock failed: ", err)
+			slog.DebugContext(ctx, "file storage unlock failed", "error", err)
 			return
 		}
-		logx.Println(ctx, "[debug] file storage unlock success")
+		slog.DebugContext(ctx, "file storage unlock success")
 	}()
 	if err := s.restore(ctx); err != nil {
 		return err
 	}
 	if err := fn(ctx); err != nil {
-		logx.Println(ctx, "[debug] transactional function failed:", err)
+		slog.DebugContext(ctx, "transactional function failed", "error", err)
 		return err
 	}
 	if err := s.store(ctx); err != nil {
 		return err
 	}
-	logx.Println(ctx, "[debug] file storage store success")
+	slog.DebugContext(ctx, "file storage store success")
 	return nil
 }
 
 func (s *FileStorage) restore(ctx context.Context) error {
 	fp, err := os.Open(s.FilePath)
 	if err != nil {
-		logx.Println(ctx, "[warn] failed restore failed:", err)
+		slog.WarnContext(ctx, "failed restore", "error", err)
 		return nil
 	}
 	defer fp.Close()
 	decoder := gob.NewDecoder(fp)
 	if err := decoder.Decode(s); err != nil && err != io.EOF {
-		log.Printf("[error] failed restore file storage: %s", err.Error())
+		slog.ErrorContext(ctx, "failed restore file storage", "error", err)
 		return err
 	}
 	return nil
@@ -572,15 +596,15 @@ func (s *FileStorage) restore(ctx context.Context) error {
 func (s *FileStorage) store(ctx context.Context) error {
 	fp, err := os.Create(s.FilePath)
 	if err != nil {
-		logx.Printf(ctx, "[error] failed store to file storage: create file: %s", err.Error())
+		slog.ErrorContext(ctx, "failed store to file storage: create file", "error", err)
 		return err
 	}
 	defer fp.Close()
 	encoder := gob.NewEncoder(fp)
 	if err := encoder.Encode(s); err != nil {
-		logx.Printf(ctx, "[error] failed store to file storage: encode gob: %s", err.Error())
+		slog.ErrorContext(ctx, "failed store to file storage: encode gob", "error", err)
 		return err
 	}
-	log.Printf("[debug] file storage store to `%s`", s.FilePath)
+	slog.DebugContext(ctx, "file storage store", "file_path", s.FilePath)
 	return nil
 }
